requirements.txt
absl-py==0.8.0
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
gast==0.3.2
greenlet==0.4.15
grpcio==1.24.0
h5py==2.10.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter-client==5.3.3
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lmdb==0.97
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mock==3.0.5
msgpack==0.5.6
msgpack-numpy==0.4.4.3
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.2
open3d==0.8.0.0
open3d-python==0.7.0.0
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.2
ptyprocess==0.6.0
pyarrow==0.14.1
Pygments==2.4.2
pynvim==0.4.1
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
PyYAML==5.3.1
pyzmq==18.1.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.12.0
tabulate==0.8.5
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorpack==0.8.9
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
